+++ packages/query-core/src/query.ts
@@ -208,10 +208,9 @@ export class Query<
     if (this.state && this.state.data === undefined) {
       const defaultState = getDefaultState(this.options)
       if (defaultState.data !== undefined) {
-        this.setData(defaultState.data, {
-          updatedAt: defaultState.dataUpdatedAt,
-          manual: true,
-        })
+        this.setState(
+          successState(defaultState.data, defaultState.dataUpdatedAt),
+        )
         this.#initialState = defaultState
       }
     }
@@ -633,12 +632,8 @@ export class Query<
         case 'success':
           const newState = {
             ...state,
-            data: action.data,
+            ...successState(action.data, action.dataUpdatedAt),
             dataUpdateCount: state.dataUpdateCount + 1,
-            dataUpdatedAt: action.dataUpdatedAt ?? Date.now(),
-            error: null,
-            isInvalidated: false,
-            status: 'success' as const,
             ...(!action.manual && {
               fetchStatus: 'idle' as const,
               fetchFailureCount: 0,
@@ -661,6 +656,9 @@ export class Query<
             fetchFailureReason: error,
             fetchStatus: 'idle',
             status: 'error',
+            // flag existing data as invalidated if we get a background error
+            // note that "no data" always means stale so we can set unconditionally here
+            isInvalidated: true,
           }
         case 'invalidate':
           return {
@@ -706,6 +704,16 @@ export function fetchState<
         status: 'pending',
       } as const)),
   } as const
+}
+
+function successState<TData>(data: TData | undefined, dataUpdatedAt?: number) {
+  return {
+    data,
+    dataUpdatedAt: dataUpdatedAt ?? Date.now(),
+    error: null,
+    isInvalidated: false,
+    status: 'success' as const,
+  }
 }
 
 function getDefaultState<