+++ packages/query-core/src/types.ts
@@ -323,9 +323,9 @@ export interface QueryObserverOptions<
   TQueryKey extends QueryKey = QueryKey,
   TPageParam = never,
 > extends WithRequired<
-    QueryOptions<TQueryFnData, TError, TQueryData, TQueryKey, TPageParam>,
-    'queryKey'
-  > {
+  QueryOptions<TQueryFnData, TError, TQueryData, TQueryKey, TPageParam>,
+  'queryKey'
+> {
   /**
    * Set this to `false` or a function that returns `false` to disable automatic refetching when the query mounts or changes query keys.
    * To refetch the query, use the `refetch` method returned from the `useQuery` instance.
@@ -467,7 +467,9 @@ export interface InfiniteQueryObserverOptions<
   TData = TQueryFnData,
   TQueryKey extends QueryKey = QueryKey,
   TPageParam = unknown,
-> extends QueryObserverOptions<
+>
+  extends
+    QueryObserverOptions<
       TQueryFnData,
       TError,
       TData,
@@ -501,9 +503,9 @@ export interface FetchQueryOptions<
   TQueryKey extends QueryKey = QueryKey,
   TPageParam = never,
 > extends WithRequired<
-    QueryOptions<TQueryFnData, TError, TData, TQueryKey, TPageParam>,
-    'queryKey'
-  > {
+  QueryOptions<TQueryFnData, TError, TData, TQueryKey, TPageParam>,
+  'queryKey'
+> {
   initialPageParam?: never
   /**
    * The time in milliseconds after data is considered stale.
@@ -519,12 +521,12 @@ export interface EnsureQueryDataOptions<
   TQueryKey extends QueryKey = QueryKey,
   TPageParam = never,
 > extends FetchQueryOptions<
-    TQueryFnData,
-    TError,
-    TData,
-    TQueryKey,
-    TPageParam
-  > {
+  TQueryFnData,
+  TError,
+  TData,
+  TQueryKey,
+  TPageParam
+> {
   revalidateIfStale?: boolean
 }
 
@@ -585,13 +587,15 @@ export interface RefetchOptions extends ResultOptions {
   cancelRefetch?: boolean
 }
 
-export interface InvalidateQueryFilters<TQueryKey extends QueryKey = QueryKey>
-  extends QueryFilters<TQueryKey> {
+export interface InvalidateQueryFilters<
+  TQueryKey extends QueryKey = QueryKey,
+> extends QueryFilters<TQueryKey> {
   refetchType?: QueryTypeFilter | 'none'
 }
 
-export interface RefetchQueryFilters<TQueryKey extends QueryKey = QueryKey>
-  extends QueryFilters<TQueryKey> {}
+export interface RefetchQueryFilters<
+  TQueryKey extends QueryKey = QueryKey,
+> extends QueryFilters<TQueryKey> {}
 
 export interface InvalidateOptions extends RefetchOptions {}
 export interface ResetOptions extends RefetchOptions {}
@@ -1076,7 +1080,7 @@ export type InfiniteQueryObserverResult<
 export type MutationKey = Register extends {
   mutationKey: infer TMutationKey
 }
-  ? TMutationKey extends Array<unknown>
+  ? TMutationKey extends ReadonlyArray<unknown>
     ? TMutationKey
     : TMutationKey extends Array<unknown>
       ? TMutationKey
@@ -1265,11 +1269,11 @@ export interface MutationObserverIdleResult<
   TVariables = void,
   TOnMutateResult = unknown,
 > extends MutationObserverBaseResult<
-    TData,
-    TError,
-    TVariables,
-    TOnMutateResult
-  > {
+  TData,
+  TError,
+  TVariables,
+  TOnMutateResult
+> {
   data: undefined
   variables: undefined
   error: null
@@ -1286,11 +1290,11 @@ export interface MutationObserverLoadingResult<
   TVariables = void,
   TOnMutateResult = unknown,
 > extends MutationObserverBaseResult<
-    TData,
-    TError,
-    TVariables,
-    TOnMutateResult
-  > {
+  TData,
+  TError,
+  TVariables,
+  TOnMutateResult
+> {
   data: undefined
   variables: TVariables
   error: null
@@ -1307,11 +1311,11 @@ export interface MutationObserverErrorResult<
   TVariables = void,
   TOnMutateResult = unknown,
 > extends MutationObserverBaseResult<
-    TData,
-    TError,
-    TVariables,
-    TOnMutateResult
-  > {
+  TData,
+  TError,
+  TVariables,
+  TOnMutateResult
+> {
   data: undefined
   error: TError
   variables: TVariables
@@ -1328,11 +1332,11 @@ export interface MutationObserverSuccessResult<
   TVariables = void,
   TOnMutateResult = unknown,
 > extends MutationObserverBaseResult<
-    TData,
-    TError,
-    TVariables,
-    TOnMutateResult
-  > {
+  TData,
+  TError,
+  TVariables,
+  TOnMutateResult
+> {
   data: TData
   error: null
   variables: TVariables