+++ packages/react-query-devtools/eslint.config.js
@@ -1,24 +1,22 @@
 // @ts-check
 
 import pluginReact from '@eslint-react/eslint-plugin'
-import * as reactHooks from 'eslint-plugin-react-hooks'
+import reactHooks from 'eslint-plugin-react-hooks'
 import rootConfig from './root.eslint.config.js'
 
 export default [
   ...rootConfig,
-  reactHooks.configs.recommended,
+  // @ts-expect-error wtf
+  ...reactHooks.configs['recommended-latest'],
   {
     files: ['**/*.{ts,tsx}'],
     ...pluginReact.configs.recommended,
     rules: {
       '@eslint-react/no-context-provider': 'off', // We need to be React 18 compatible
-    },
-  },
-  {
-    rules: {
       'react-hooks/exhaustive-deps': 'error',
       'react-hooks/rules-of-hooks': 'error',
-      'react-hooks/react-compiler': 'error',
+      'react-hooks/unsupported-syntax': 'error',
+      'react-hooks/incompatible-library': 'error',
     },
   },
 ]