+++ packages/eslint-plugin-query/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @tanstack/eslint-plugin-query
 
+## 5.91.2
+
+### Patch Changes
+
+- fix: allow useQueries with combine property in no-unstable-deps rule ([#9720](https://github.com/TanStack/query/pull/9720))
+
+## 5.91.1
+
+### Patch Changes
+
+- avoid typescript import in no-void-query-fn rule ([#9759](https://github.com/TanStack/query/pull/9759))
+
 ## 5.91.0
 
 ### Minor Changes